Refactor marshalling attribute parsing into its own type to enable better holistic analysis of all use-site marshalling attributes.
